Add files via upload
This project aims to provide a user-friendly and visually appealing signup form that allows users to register for an account on the platform. Leveraging Flask's powerful back-end capabilities, the project securely stores user information in a database, ensuring data privacy. The frontend, designed with HTML and Bootstrap, offers a seamless and responsive experience across devices.